blindsided by the singlesource, put templates in the header
