+++ code/espurna/web.h
@@ -44,6 +44,7 @@ struct AsyncWebPrint : public Print {
 
     // To be able to safely output data right from the request callback,
     // we schedule a 'printer' task that will print into the request response buffer via AsyncChunkedResponse
+    // Note: implementation must be included in the header
     template<typename CallbackType>
     static void scheduleFromRequest(const AsyncWebPrintConfig& config, AsyncWebServerRequest*, CallbackType);
 
@@ -81,6 +82,9 @@ struct AsyncWebPrint : public Print {
     void _prepareRequest();
 
 };
+
+// Implementation of tempated constructors
+#include "web_asyncwebprint.h"
 
 using web_body_callback_f = std::function<bool(AsyncWebServerRequest*, uint8_t* data, size_t len, size_t index, size_t total)>;
 using web_request_callback_f = std::function<bool(AsyncWebServerRequest*)>;
